requirements.txt
fastapi==0.120.2
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
anthropic==0.72.0
langgraph==1.0.2
langchain==1.0.3
langchain-anthropic==1.0.0
replicate==1.0.7
pydantic==2.12.3
pydantic-settings==2.1.0
dependency-injector==4.42.0
httpx==0.28.1